+++ tests/specs/solver.spec.js
@@ -1,10 +1,8 @@
 import expect from '../fixtures/mocha_proxy.fixt';
 import {
-  fixt_arrdom_empty,
   fixt_arrdom_nums,
   fixt_arrdom_range,
   fixt_arrdom_ranges,
-  fixt_dom_nums,
   fixt_dom_ranges,
   stripAnonVarsFromArrays,
 } from '../fixtures/domain.fixt';
@@ -19,7 +17,6 @@ import {
   LOG_SOLVES,
   LOG_MAX,
   LOG_MIN,
-  NO_SUCH_VALUE,
   SUB,
   SUP,
 } from '../../src/helpers';
@@ -977,15 +974,6 @@ describe('solver.spec', function() {
       });
     });
 
-    describe('solver.space_add_var_range', function() {
-
-      it('should just map to config_addVarRange', function() {
-        let solver = new Solver();
-
-        expect(solver.space_add_var_range('foo', 1, 2)).to.equal('foo');
-      });
-    });
-
     describe('solver.domain_fromList', function() {
 
       it('should map to domain_fromList', function() {
@@ -998,30 +986,6 @@ describe('solver.spec', function() {
         let solver = new Solver();
 
         expect(solver.domain_fromList([1, 2, 4, 5, 7, 9, 10, 11, 12, 13, 15])).to.eql([1, 2, 4, 5, 7, 7, 9, 13, 15, 15]);
-      });
-    });
-
-    describe('solver.domain_max', function() {
-
-      it('should work on a domain', function() {
-        let solver = new Solver();
-
-        expect(solver.domain_max(fixt_arrdom_nums(0, 1, 4, 6))).to.eql(6);
-      });
-
-      it('should return NO_SUCH_VALUE if the domain is empty', function() {
-        let solver = new Solver();
-
-        expect(solver.domain_max(fixt_arrdom_empty())).to.eql(NO_SUCH_VALUE);
-      });
-    });
-
-    describe('solver.domain_toList', function() {
-
-      it('should return an array of values for given domain', function() {
-        let solver = new Solver();
-
-        expect(solver.domain_toList(fixt_dom_nums(0, 1, 2, 3, 8, 10))).to.eql([0, 1, 2, 3, 8, 10]);
       });
     });
   });